refactor : ThemeDao
