+++ app-client/src/main/java/com/welcomeToJeju/moj/handler/ranking/ThemeRankingHandler.java
@@ -1,15 +1,15 @@
-package com.welcomeToJeju.moj.handler.rank;
+package com.welcomeToJeju.moj.handler.ranking;
 
 import com.welcomeToJeju.moj.dao.ThemeDao;
 import com.welcomeToJeju.moj.domain.Theme;
 import com.welcomeToJeju.moj.handler.Command;
 import com.welcomeToJeju.moj.handler.CommandRequest;
 
-public class ThemeRankHandler implements Command {
+public class ThemeRankingHandler implements Command {
 
   ThemeDao themeDao;
 
-  public ThemeRankHandler(ThemeDao themeDao) {
+  public ThemeRankingHandler(ThemeDao themeDao) {
     this.themeDao = themeDao;
   }
 
@@ -18,7 +18,7 @@ public class ThemeRankHandler implements Command {
     System.out.println("[테마 순위 보기]");
 
     int no = 1;
-    for(Theme theme : themeDao.sortThemeByViewCount()) {
+    for(Theme theme : themeDao.themeRankingByViewCount()) {
       System.out.printf("%d위 > %s (조회수 : %d)\n",
           no++, theme.getTitle(), theme.getViewCount());
     }
